pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "score-vision"
version = "0.3.0"
description = "Score Vision - Soccer Video Analysis Framework"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Score Vision Team", email = "hello@wearescore.com"}
]

[project.optional-dependencies]
miner = [
    "supervision>=0.16.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "transformers>=4.35.0",
    "umap-learn==0.5.7",
    "scikit-learn>=1.3.0",
    "tqdm>=4.66.0",
    "sentencepiece>=0.1.99",
    "protobuf>=4.24.0",
    "llvmlite>=0.41.0",
    "fastapi>=0.104.0",
    "ultralytics>=8.0.0",
    "python-multipart>=0.0.6",
    "uvicorn>=0.24.0",
    "gdown>=5.2.0",
    "python-dotenv>=1.0.1",
    "huggingface-hub>=0.19.0",
    "loguru>=0.7.2",
    "aiohttp>=3.9.1",
    "httpx>=0.25.2",
    "pydantic>=2.5.2",
    "PyYAML>=6.0.1",
    "bt-decode==0.5.0a2",
    "async-substrate-interface==1.0.0",
    "fiber[full] @ git+https://github.com/rayonlabs/fiber.git@2.3.0#egg=fiber"
]
validator = [
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.2",
    "PyYAML>=6.0.1",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.1",
    "aiohttp>=3.9.1",
    "httpx>=0.25.2",
    "loguru>=0.7.2",
    "substrate-interface>=1.7.4",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "openai>=1.0.0",
    "sqlalchemy>=2.0.23",
    "tenacity>=8.2.3",
    "pytz>=2023.3",
    "bt-decode>=0.5.0a2",
    "open_clip_torch>=2.23.0",
    "transformers>=4.36.2",
    "torch>=2.2.0",
    "async-substrate-interface>=1.0.0",
    "transformers>=4.34.0",
    "accelerate>=0.24.0",
    "fiber[full] @ git+https://github.com/rayonlabs/fiber.git@2.5.0#egg=fiber",
    "zstandard==0.23.0"
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]

[tool.setuptools]
packages = ["miner", "validator"]
